[bug-fix] 나만보기 및 아카이브가 없었던 것들도 포함하여 나오게 설정

+++ kilometer-backend-domain/src/main/java/com/kilometer/domain/search/additionalinfo/ListItemAdditionalInfo.java
@@ -9,6 +9,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class ListItemAdditionalInfo {
     private int heartCount;
-    private float grade;
-    private float archiveCount;
+    private double grade;
+    private long archiveCount;
 }
